tstest: rename from testy.
